requirements.txt
asgiref==3.3.4
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.9.1
bidict==0.21.2
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
constantly==15.1.0
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.1
gunicorn==20.0.4
hyperlink==21.0.0
idna==2.9
incremental==21.3.0
lxml==4.5.1
psycopg2==2.8.6
pycparser==2.20
python-dotenv==0.13.0
python-engineio==4.2.1
python-socketio==5.4.0
pytube3==9.6.4
pytz==2020.1
requests==2.24.0
six==1.16.0
soupsieve==2.0.1
sqlparse==0.4.1
twisted-iocpsupport==1.0.1
typing-extensions==3.7.4.3
urllib3==1.25.9
whitenoise==5.2.0
zope.interface==5.4.0
